chore: moved is* for worship classifications to worship administrative unit model
